Bird's Movement Update, Floor and Pipes Implemented
Updated the bird's movement so that it feels better, and also added to the scene pipes (which are automatically generated by a corroutine) and floor, which must still be polished. Finally, added a script that moves to the left every object in each layer separately (will be used to achieve parallax effect later on).